watermarking/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
# pytorch-wavelets removed: replaced by models/haar_dwt.py (pure PyTorch, no external dep)
pytorch-msssim>=1.0.0
lpips>=0.1.4
kornia>=0.7.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
